chore: split tests in multiple files
