+++ tests/TailwindVariants.NET.Tests/Models/ButtonSlots.cs
@@ -1,0 +1,15 @@
+﻿namespace TailwindVariants.NET.Tests;
+
+#pragma warning disable CS0436
+
+public partial class ButtonSlots : ISlots
+{
+    [Slot("root")]
+    public string? Base { get; set; }
+
+    public string? Icon { get; set; }
+
+    public string? Label { get; set; }
+}
+
+#pragma warning restore CS0436
