+++ tests/TailwindVariants.NET.Tests/Models/GhostButtonComponent.cs
@@ -1,0 +1,8 @@
+﻿namespace TailwindVariants.NET.Tests;
+
+public partial class GhostButtonComponent : ISlotted<GhostButtonSlots>
+{
+    public string? Class { get; set; }
+    public GhostButtonSlots? Classes { get; set; }
+    public string? Variant { get; set; }
+}
