+++ tests/TailwindVariants.NET.Tests/Models/GhostButtonSlots.cs
@@ -1,0 +1,6 @@
+﻿namespace TailwindVariants.NET.Tests;
+
+public partial class GhostButtonSlots : ButtonSlots
+{
+    public string? Overlay { get; set; }
+}
